app/requirements.txt
pandas
requests
SQLAlchemy
psycopg2-binary
python-dotenv
